Add getOne method to TodoEntityService and handle NotFoundException in TodosService

+++ src/todos/todos.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { TodoDto } from './dto/todo.dto';
 import { TodoEntityService } from './todo-entity.service';
 import {
@@ -20,10 +20,18 @@ export class TodosService {
   }
 
   async update(data: UpdateTodoInput): Promise<TodoDto> {
+    const todo = await this.todoEntityService.getOne(data.id, data.user_id);
+    if (!todo) {
+      throw new NotFoundException();
+    }
     return this.todoEntityService.update(data);
   }
 
   async delete(data: DeleteTodoInput): Promise<boolean> {
+    const todo = await this.todoEntityService.getOne(data.id, data.user_id);
+    if (!todo) {
+      throw new NotFoundException();
+    }
     return this.todoEntityService.delete(data);
   }
 }
